Move useMutation defer tests to own file
